+++ internal/domain/wallet/repository/wallets.go
@@ -6,19 +6,20 @@ import (
 	"time"
 
 	"personal-finance/internal/model"
+	"personal-finance/internal/plataform/authentication"
 
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
 
 type Repository interface {
-	RecalculateBalance(ctx context.Context, walletID *uuid.UUID, userID string) error
-	Add(ctx context.Context, wallet model.Wallet, userID string) (model.Wallet, error)
-	FindAll(ctx context.Context, userID string) ([]model.Wallet, error)
-	FindByID(ctx context.Context, id *uuid.UUID, userID string) (model.Wallet, error)
-	Update(ctx context.Context, id *uuid.UUID, wallet model.Wallet, userID string) (model.Wallet, error)
+	RecalculateBalance(ctx context.Context, walletID *uuid.UUID) error
+	Add(ctx context.Context, wallet model.Wallet) (model.Wallet, error)
+	FindAll(ctx context.Context) ([]model.Wallet, error)
+	FindByID(ctx context.Context, id *uuid.UUID) (model.Wallet, error)
+	Update(ctx context.Context, id *uuid.UUID, wallet model.Wallet) (model.Wallet, error)
 	Delete(ctx context.Context, id *uuid.UUID) error
-	UpdateConsistent(_ context.Context, tx *gorm.DB, wallet model.Wallet, userID string) (model.Wallet, error)
+	UpdateConsistent(_ context.Context, tx *gorm.DB, wallet model.Wallet) (model.Wallet, error)
 }
 
 type PgRepository struct {
@@ -29,9 +30,10 @@ func NewPgRepository(gorm *gorm.DB) Repository {
 	return PgRepository{Gorm: gorm}
 }
 
-func (p PgRepository) RecalculateBalance(ctx context.Context, walletID *uuid.UUID, userID string) error {
+func (p PgRepository) RecalculateBalance(ctx context.Context, walletID *uuid.UUID) error {
+	userID := ctx.Value(authentication.UserID).(string)
 	var recalculatedBalance float64
-	wallet, err := p.FindByID(ctx, walletID, userID)
+	wallet, err := p.FindByID(ctx, walletID)
 	if err != nil {
 		return err
 	}
@@ -49,7 +51,7 @@ func (p PgRepository) RecalculateBalance(ctx context.Context, walletID *uuid.UUI
 	}
 
 	wallet.Balance = wallet.InitialBalance + recalculatedBalance
-	_, err = p.Update(ctx, walletID, wallet, userID)
+	_, err = p.Update(ctx, walletID, wallet)
 	if err != nil {
 		return err
 	}
@@ -57,7 +59,8 @@ func (p PgRepository) RecalculateBalance(ctx context.Context, walletID *uuid.UUI
 	return nil
 }
 
-func (p PgRepository) Add(_ context.Context, wallet model.Wallet, userID string) (model.Wallet, error) {
+func (p PgRepository) Add(ctx context.Context, wallet model.Wallet) (model.Wallet, error) {
+	userID := ctx.Value(authentication.UserID).(string)
 	now := time.Now()
 	id := uuid.New()
 
@@ -76,7 +79,8 @@ func (p PgRepository) Add(_ context.Context, wallet model.Wallet, userID string)
 	return wallet, nil
 }
 
-func (p PgRepository) FindAll(_ context.Context, userID string) ([]model.Wallet, error) {
+func (p PgRepository) FindAll(ctx context.Context) ([]model.Wallet, error) {
+	userID := ctx.Value(authentication.UserID).(string)
 	var wallets []model.Wallet
 	result := p.Gorm.Where("user_id=?", userID).Order("description").Find(&wallets)
 	if err := result.Error; err != nil {
@@ -85,8 +89,9 @@ func (p PgRepository) FindAll(_ context.Context, userID string) ([]model.Wallet,
 	return wallets, nil
 }
 
-func (p PgRepository) FindByID(_ context.Context, id *uuid.UUID, userID string) (model.Wallet, error) {
+func (p PgRepository) FindByID(ctx context.Context, id *uuid.UUID) (model.Wallet, error) {
 	var wallet model.Wallet
+	userID := ctx.Value(authentication.UserID).(string)
 	result := p.Gorm.Where("user_id=?", userID).First(&wallet, id)
 	if err := result.Error; err != nil {
 		return model.Wallet{}, err
@@ -94,8 +99,8 @@ func (p PgRepository) FindByID(_ context.Context, id *uuid.UUID, userID string) 
 	return wallet, nil
 }
 
-func (p PgRepository) Update(_ context.Context, id *uuid.UUID, wallet model.Wallet, userID string) (model.Wallet, error) {
-	w, err := p.FindByID(context.Background(), id, userID)
+func (p PgRepository) Update(ctx context.Context, id *uuid.UUID, wallet model.Wallet) (model.Wallet, error) {
+	w, err := p.FindByID(ctx, id)
 	if err != nil {
 		return model.Wallet{}, err
 	}
@@ -118,7 +123,7 @@ func (p PgRepository) Update(_ context.Context, id *uuid.UUID, wallet model.Wall
 	result := p.Gorm.Save(&w)
 
 	if shouldRecalculate {
-		if err := p.RecalculateBalance(context.Background(), id, userID); err != nil {
+		if err := p.RecalculateBalance(ctx, id); err != nil {
 			return model.Wallet{}, err
 		}
 	}
@@ -136,8 +141,8 @@ func (p PgRepository) Delete(_ context.Context, id *uuid.UUID) error {
 	return nil
 }
 
-func (p PgRepository) UpdateConsistent(_ context.Context, tx *gorm.DB, wallet model.Wallet, userID string) (model.Wallet, error) {
-	w, err := p.FindByID(context.Background(), wallet.ID, userID)
+func (p PgRepository) UpdateConsistent(ctx context.Context, tx *gorm.DB, wallet model.Wallet) (model.Wallet, error) {
+	w, err := p.FindByID(ctx, wallet.ID)
 	if err != nil {
 		return model.Wallet{}, err
 	}